MplSearchResultProductPopulator.java changes for sonar fix

+++ ext/marketplace/marketplacefacades/src/com/tisl/mpl/search/converters/populator/MplSearchResultProductPopulator.java
@@ -226,7 +226,7 @@ public class MplSearchResultProductPopulator extends SearchResultVariantProductP
 			final double savingsAmt = mrpPriceValue.doubleValue() - priceValue.doubleValue();
 			final double calculatedPerSavings = (savingsAmt / mrpPriceValue.doubleValue()) * 100;
 			final double roundedOffValue = Math.round(calculatedPerSavings * 100.0) / 100.0;
-			final PriceData priceData = getPriceDataFactory().create(PriceDataType.BUY, BigDecimal.valueOf(roundedOffValue),
+			final PriceData priceData = getPriceDataFactory().create(PriceDataType.BUY, BigDecimal.valueOf(Double.valueOf(roundedOffValue).intValue()),
 					getCommonI18NService().getCurrentCurrency());
 
 			target.setSavingsOnProduct(priceData);
